feat: implement route generation utilities for Next.js and Vite; add dynamic segment handling and refactor existing functions
