+++ src/templates/vite-react/route-generator.ts
@@ -4,57 +4,15 @@ import { getCalmUiJson } from "../../lib/get-calmui-json";
 import textToCamelCase from "../../utils/text-to-camel-case";
 import chalk from "chalk";
 import { upperCamelCase } from "../../utils/text-to-upper-camel-case";
+import { staticParentPath } from "./utils/static-parent-path";
+import { lastStaticSegment } from "./utils/last-static-segment";
+import { routeToTanstackPath } from "./utils/route-to-tanstack-path";
+import { getTanstackDynamicSegments } from "./utils/get-tanstack-dynamic-segments";
 
 const successLog = (text: string) => console.log(chalk.greenBright(text));
 const infoLog = (text: string) => console.log(chalk.blueBright(text));
 
 const cwd = process.cwd();
-
-// ? Helpers only for route generation with vite and tanstack
-
-/**
- * Converts a route string to the TanStack Router format by replacing dynamic segments with the "$" prefix.
- * @param route - The route pattern (e.g., "/folder1/:folder2/:folder3")
- * @returns The TanStack Router format (e.g., "/folder1/$folder2/$folder3")
- */
-function routeToTanstackPath(route: string) {
-  return route
-    .split("/")
-    .map((part) => (part.startsWith(":") ? `$${part.slice(1)}` : part))
-    .join("/");
-}
-
-/**
- * Extracts the last static segment from a route string.
- *
- * @param route - The route pattern (e.g., "/folder1/:folder2/:folder3")
- * @returns The last static segment (e.g., "folder3")
- */
-function lastStaticSegment(route: string) {
-  const parts = route.split("/");
-  let last = "";
-  for (const part of parts) {
-    if (part.startsWith(":")) break;
-    if (part) last = part;
-  }
-  return last;
-}
-
-/**
- * Extracts the parent static segments from a route string.
- *
- * @param route - The route pattern (e.g., "/folder1/:folder2/:folder3")
- * @returns The parent static segments joined by "/" (e.g., "folder1")
- */
-function staticParentPath(route: string) {
-  const parts = route.split("/");
-  let staticParts = [];
-  for (const part of parts) {
-    if (part.startsWith(":")) break;
-    if (part) staticParts.push(part);
-  }
-  return staticParts.join("/");
-}
 
 // -------------------- FUNCTIONS --------------------
 // 1. Route generation
@@ -73,6 +31,7 @@ const generateRoute = ({
 }) => {
   const tanstackPath = routeToTanstackPath(route);
   const routePath = path.join(cwd, "src", "routes", tanstackPath);
+  const { isDynamic, segments } = getTanstackDynamicSegments(route);
   fs.ensureDirSync(routePath);
   const newRoutePath = `${routePath}/index.${fileExtension}x`;
   const isRouteExists = fs.existsSync(newRoutePath);
@@ -80,13 +39,22 @@ const generateRoute = ({
     fs.ensureFileSync(newRoutePath);
     fs.writeFileSync(
       newRoutePath,
-      `import { createFileRoute } from '@tanstack/react-router'
+      `import { createFileRoute, ${
+        isDynamic ? "useParams" : ""
+      } } from '@tanstack/react-router'
 
 export const Route = createFileRoute('${tanstackPath}')({
   component: RouteComponent,
 })
 
 function RouteComponent() {
+  ${
+    isDynamic
+      ? `const { ${segments.join(", ")} } = useParams({
+    from: '${tanstackPath}/',
+  });`
+      : ""
+  }
   return <div>Hello "${route}"!</div>
 }
 `
